+++ CHANGELOG.md
@@ -43,6 +43,7 @@
       var d = 0.1+0.2
       assert "$d".toFloat()! == $d
       ```
+- add ``help`` attribute param to ``Arg`` attribute. now specify detail message of positional argument
 
 #### Builtin